+++ src/test/kotlin/graphql/routing/QueriesAndMutationsTest.kt
@@ -5,11 +5,10 @@ import com.fasterxml.jackson.module.kotlin.readValue
 import com.neelkamath.omniChat.*
 import com.neelkamath.omniChat.db.tables.GroupChats
 import com.neelkamath.omniChat.db.tables.create
-import com.neelkamath.omniChat.graphql.operations.ONLINE_STATUS_FRAGMENT
 import com.neelkamath.omniChat.graphql.operations.READ_ACCOUNT_QUERY
 import com.neelkamath.omniChat.graphql.operations.STARRED_MESSAGE_FRAGMENT
+import com.neelkamath.omniChat.graphql.operations.TYPING_STATUS_FRAGMENT
 import com.neelkamath.omniChat.graphql.operations.UPDATE_GROUP_CHAT_TITLE_QUERY
-
 import io.ktor.application.*
 import io.ktor.http.*
 import io.ktor.server.testing.*
@@ -41,9 +40,9 @@ class QueriesAndMutationsTest {
                             }
                         }
                         
-                        query ReadOnlineStatuses {
-                            readOnlineStatuses {
-                                $ONLINE_STATUS_FRAGMENT
+                        query ReadTypingStatuses {
+                            readTypingStatuses {
+                                $TYPING_STATUS_FRAGMENT
                             }
                         }
                     """